# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "astrea-sda-api",
    "description": "A high-performance Rust API for Space Domain Awareness (SDA), satellite tracking, SGP4 propagation, and visibility prediction",
    "contact": {
      "name": "Michael Wedd"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Developer Account Login & JWT Token Generation",
        "description": "Authenticates developer account credentials and returns an RS256/Supabase JWT Bearer token for API access and Swagger UI authorization.",
        "operationId": "loginHandler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authenticated successfully and JWT token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Current Authenticated Developer Profile",
        "description": "Returns decoded JWT claims and profile info for the authenticated caller.",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Decoded caller JWT claims",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Claims"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Developer Account Registration",
        "description": "Registers a new developer account and issues an authenticated RS256/Supabase JWT Bearer token.",
        "operationId": "signupHandler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Developer account registered and JWT token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid registration request or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "RFC 7523 Private Key JWT Client Assertion M2M Token Exchange",
        "description": "Authenticates machine-to-machine service accounts via signed RSA Private Key assertions (`client_assertion_type=urn:ietf:params:oauth:client-assertion-type:jwt-bearer`) and returns a scoped RS256 JWT Bearer token.",
        "operationId": "tokenExchangeHandler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientAssertionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "M2M Client assertion verified and JWT Bearer token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid client assertion format or unsupported grant type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid assertion signature or expired timestamp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/conjunctions/collision-probability": {
      "post": {
        "tags": [
          "Astrodynamics"
        ],
        "summary": "Calculate Conjunction Collision Probability (Pc)",
        "description": "Computes 2D encounter-plane collision probability (Pc) using Foster's algorithm given miss distance, relative velocity, hard-body radius, and combined position covariance uncertainty. Protected by JWT auth (requires 'viewer', 'editor', or 'admin' role).",
        "operationId": "calculateCollisionProbability",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollisionProbabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Collision probability computed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollisionProbabilityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/conjunctions/search": {
      "get": {
        "tags": [
          "Astrodynamics"
        ],
        "summary": "Search Conjunctions & Satellite Collision Radar",
        "description": "Evaluates close-approach orbital conjunctions across all active satellites using Rayon parallel execution. Protected by JWT auth (requires 'viewer', 'editor', or 'admin' role).",
        "operationId": "searchConjunctions",
        "parameters": [
          {
            "name": "max_distance_km",
            "in": "query",
            "description": "Maximum threshold distance in km between satellites to trigger a conjunction match (default: 10.0 km)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            }
          },
          {
            "name": "duration_hours",
            "in": "query",
            "description": "Forecast window forward in hours (default: 24 hrs, max: 72)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conjunction radar search completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConjunctionSearchResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal calculation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/pipelines/sync": {
      "post": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Trigger CelesTrak Pipeline Sync",
        "description": "Triggers automated CelesTrak discovery pipeline synchronization for a specific satellite group. Protected by JWT auth (requires 'admin' role).",
        "operationId": "triggerPipelineSync",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "description": "CelesTrak satellite group name to synchronize. Options: 'curated' (default: stations, visual, last-30-days), 'all', 'stations', 'visual', 'starlink', 'weather', 'last-30-days', 'active', or a comma-separated list (e.g. 'stations,visual').",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "max_age_hours",
            "in": "query",
            "description": "Maximum allowed data age in hours before triggering a CelesTrak query (default: 12.0 hours). If existing data was synced within this period, CelesTrak requests are skipped to prevent spamming.",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "Force an immediate fetch from CelesTrak, bypassing freshness checks.",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline sync completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineSyncResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid or unknown satellite group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient role permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Pipeline sync execution failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/satellites": {
      "get": {
        "tags": [
          "Satellites"
        ],
        "summary": "List Satellites",
        "description": "Retrieves a paginated list of satellites using base64 checkpoint cursors. Protected by JWT auth (requires 'viewer', 'editor', or 'admin' role).",
        "operationId": "listSatellites",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of satellites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseSatellite"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Satellites"
        ],
        "summary": "Create Satellite",
        "description": "Creates a new satellite record from Two-Line Element (TLE) set data. Protected by JWT auth (requires 'editor' or 'admin' role).",
        "operationId": "createSatellite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSatelliteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Satellite created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Satellite"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient role permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/satellites/decay-watch": {
      "get": {
        "tags": [
          "Astrodynamics"
        ],
        "summary": "Atmospheric Drag & Orbital Decay Re-Entry Watch",
        "description": "Scans the satellite catalog for debris and satellites experiencing severe atmospheric drag or nearing uncontrolled atmospheric re-entry. Protected by JWT auth (requires 'viewer', 'editor', or 'admin' role).",
        "operationId": "getDecayWatch",
        "parameters": [
          {
            "name": "max_perigee_km",
            "in": "query",
            "description": "Maximum perigee altitude threshold in km to flag re-entry hazard (default: 300.0 km)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            }
          },
          {
            "name": "min_bstar",
            "in": "query",
            "description": "Minimum B* atmospheric drag coefficient threshold (default: 0.0001)",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of decaying objects to return (default: 50, max: 500)",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Decay watch scan completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecayWatchResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal calculation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/satellites/overhead": {
      "get": {
        "tags": [
          "Satellites"
        ],
        "summary": "Get Overhead Satellite",
        "description": "Computes the satellite closest to overhead (highest elevation) across all tracked satellites using parallel Rayon propagation. Protected by JWT auth (requires 'viewer', 'editor', or 'admin' role).",
        "operationId": "getOverheadSatellite",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "Observer latitude in decimal degrees (-90.0 to 90.0)",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "description": "Observer longitude in decimal degrees (-180.0 to 180.0)",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "alt",
            "in": "query",
            "description": "Observer altitude above sea level in meters",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "nullable": true
            }
          },
          {
            "name": "time",
            "in": "query",
            "description": "UTC timestamp for calculation epoch (defaults to current time if omitted)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Overhead satellite computed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverheadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No satellite overhead found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/satellites/{id}/decay-risk": {
      "get": {
        "tags": [
          "Satellites"
        ],
        "summary": "Get Satellite Atmospheric Drag & Decay Risk",
        "description": "Evaluates perigee/apogee altitude, B* atmospheric drag decay rate, remaining orbital lifetime, and uncontrolled re-entry risk score. Protected by JWT auth (requires 'viewer', 'editor', or 'admin' role).",
        "operationId": "getSatelliteDecayRisk",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Satellite unique UUID identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Satellite orbital decay risk assessed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SatelliteDecayRiskResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Satellite not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "description": "Response returned upon successful RS256 JWT token generation.",
        "required": [
          "token",
          "tokenType",
          "expiresIn",
          "claims"
        ],
        "properties": {
          "claims": {
            "$ref": "#/components/schemas/Claims"
          },
          "expiresIn": {
            "type": "integer",
            "description": "Token lifetime in seconds (e.g. 86400)",
            "minimum": 0
          },
          "token": {
            "type": "string",
            "description": "Encoded RS256 JWT Bearer token"
          },
          "tokenType": {
            "type": "string",
            "description": "Token type (always \"Bearer\")"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "status"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "Request body for developer account login.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "Developer email address or username"
          },
          "password": {
            "type": "string",
            "description": "Account password"
          }
        }
      },
...
      "PaginatedResponseSatellite": {
        "type": "object",
        "description": "Generic Paginated Response envelope",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Satellite"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}